245- Changed the organization code to dynamically change the placeholder rather than the value. Also Created the OnboardingPage.jsx in order to pass in the context to the page so that we can get the success messages.
